+++ src/is_function.ts
@@ -1,3 +1,3 @@
-/** Check if _value_ is function. */
-export const isFunction = (value: unknown): value is Function =>
+/** Check if _value_ is a function. */
+export const isFunction = <F extends Function>(value: unknown): value is F =>
   typeof value === 'function';