+++ test/throttle.spec.ts
@@ -1,23 +1,11 @@
-import { throttle, FRAME_TIME } from '../src';
+import { throttle } from '../src';
 
 describe('throttle', () => {
-  it('should delay function execution for 2 frames', (done) => {
+  it('should execute function immediately by default', (done) => {
     const now = Date.now();
     throttle(() => {
       try {
-        expect(Date.now() - now).toBeGreaterThan(30);
-        done();
-      } catch (error) {
-        done(error);
-      }
-    }, 2)();
-  });
-
-  it('should delay function execution for 2 frames by default', (done) => {
-    const now = Date.now();
-    throttle(() => {
-      try {
-        expect(Date.now() - now).toBeGreaterThan(30);
+        expect(Date.now() - now).toBeLessThan(1);
         done();
       } catch (error) {
         done(error);
@@ -25,58 +13,15 @@ describe('throttle', () => {
     })();
   });
 
-  it('should execute function with requestAnimationFrame funtion if frames number is 0', (done) => {
-    globalThis.requestAnimationFrame = jest.fn(
-      (callback: FrameRequestCallback) => (callback(2), 1)
-    );
+  it('should block other executions that goes right after first execution', () => {
+    const fakeFunction = jest.fn();
 
-    throttle(() => {
-      try {
-        expect(globalThis.requestAnimationFrame).toBeCalledTimes(1);
-        done();
-      } catch (error) {
-        done(error);
-      }
-    }, 0)();
-  });
-
-  it('should execute function with requestAnimationFrame funtion if frames number 1', (done) => {
-    globalThis.requestAnimationFrame = jest.fn(
-      (callback: FrameRequestCallback) => (callback(2), 1)
-    );
-
-    throttle(() => {
-      try {
-        expect(globalThis.requestAnimationFrame).toBeCalledTimes(1);
-        done();
-      } catch (error) {
-        done(error);
-      }
-    }, 1)();
-  });
-
-  it('should be invoked less than 2 times in multiple invocation', () => {
-    let count = 0;
-
-    const throttledFunction = throttle(() => count++);
+    const throttledFunction = throttle(fakeFunction, 2);
 
     throttledFunction();
     throttledFunction();
     throttledFunction();
-    throttledFunction();
-    throttledFunction();
-    throttledFunction();
 
-    expect(count).toBeLessThan(2);
-  });
-});
-
-describe('FRAME_TIME', () => {
-  it('should be exported', () => {
-    expect(FRAME_TIME).toBeDefined();
-  });
-
-  it('should be close to 16.67', () => {
-    expect(FRAME_TIME).toBeCloseTo(16.67);
+    expect(fakeFunction).toBeCalledTimes(1);
   });
 });